Disable rhizome option bugfix
Rhizome *needs* sdcard. If no sdcard is present on the device, then rhizome must
not be called.

A feature to disable it was already implemented, just add:
  rhizome.enable=0
to serval.conf (don't forget '\n').

But it was not read by the Android application not to call rhizome functions.

This commit disable rhizome calls and avoids to make the app crash when no
sdcard is present. It needs the modifications done on 'serval-dna' submodule.