Make serval work on Android-IBSS
See:
  https://groups.google.com/forum/#!topic/serval-project-developers/JGN00i00nFk
  http://www.thinktube.com/android-tech/46-android-wifi-ibss

iw binary comes from:
  http://www.onlyxool.net/porting-iw-tools-to-android/

Target SDK has changed from 8 to 10 (minSdkVersion is still 8). It allows to
know if an interface is up or down. The test is done only if the device
supports it (>= 9).

A variable "wifi.frequency" has been added in adhoc.conf (iw expects frequency
instead of channel).

Tested on Galaxy Nexus.